Move code and test files under subdirectories
Created subdirectories:
* handlers/notifications
* handlers/signals
* schedule/filters
* tests/containers
* tests/events
* tests/filters
