+++ CMakeLists.txt
@@ -188,13 +188,13 @@ target_sources(${PROGRAM_NAME} PRIVATE
     "${SCHETOOL_SRC}/${CONTAINERS_DIR}/weekday_container.cpp"
     "${SCHETOOL_SRC}/${HANDLERS_DIR}/gui_textures.cpp"
     "${SCHETOOL_SRC}/${HANDLERS_DIR}/input.cpp"
-    "${SCHETOOL_SRC}/${HANDLERS_DIR}/notification_handler_linux_impl.cpp"
-    "${SCHETOOL_SRC}/${HANDLERS_DIR}/notification_handler_win_impl.cpp"
-    "${SCHETOOL_SRC}/${HANDLERS_DIR}/notification_handler.cpp"
+    "${SCHETOOL_SRC}/${HANDLERS_DIR}/notifications/notification_handler_linux_impl.cpp"
+    "${SCHETOOL_SRC}/${HANDLERS_DIR}/notifications/notification_handler_win_impl.cpp"
+    "${SCHETOOL_SRC}/${HANDLERS_DIR}/notifications/notification_handler.cpp"
     "${SCHETOOL_SRC}/${HANDLERS_DIR}/render.cpp"
-    "${SCHETOOL_SRC}/${HANDLERS_DIR}/signal_listener_linux_impl.cpp"
-    "${SCHETOOL_SRC}/${HANDLERS_DIR}/signal_listener_win_impl.cpp"
-    "${SCHETOOL_SRC}/${HANDLERS_DIR}/signal_listener.cpp"
+    "${SCHETOOL_SRC}/${HANDLERS_DIR}/signals/signal_listener_linux_impl.cpp"
+    "${SCHETOOL_SRC}/${HANDLERS_DIR}/signals/signal_listener_win_impl.cpp"
+    "${SCHETOOL_SRC}/${HANDLERS_DIR}/signals/signal_listener.cpp"
     "${SCHETOOL_SRC}/${HANDLERS_DIR}/time_handler.cpp"
     "${SCHETOOL_SRC}/${HANDLERS_DIR}/window.cpp"
     "${SCHETOOL_SRC}/${INTERFACE_DIR}/autosave_popup_gui.cpp"
@@ -217,9 +217,9 @@ target_sources(${PROGRAM_NAME} PRIVATE
     "${SCHETOOL_SRC}/${IO_DIR}/schedule_io.cpp"
     "${SCHETOOL_SRC}/${IO_DIR}/textures.cpp"
     "${SCHETOOL_SRC}/${SCHEDULE_DIR}/element_base.cpp"
-    "${SCHETOOL_SRC}/${SCHEDULE_DIR}/filter.cpp"
-    "${SCHETOOL_SRC}/${SCHEDULE_DIR}/filter_group.cpp"
-    "${SCHETOOL_SRC}/${SCHEDULE_DIR}/filter_rule_base.cpp"
+    "${SCHETOOL_SRC}/${SCHEDULE_DIR}/filters/filter.cpp"
+    "${SCHETOOL_SRC}/${SCHEDULE_DIR}/filters/filter_group.cpp"
+    "${SCHETOOL_SRC}/${SCHEDULE_DIR}/filters/filter_rule_base.cpp"
     "${SCHETOOL_SRC}/${SCHEDULE_DIR}/schedule.cpp"
     "${SCHETOOL_SRC}/${SCHEDULE_DIR}/schedule_column.cpp"
     "${SCHETOOL_SRC}/${SCHEDULE_DIR}/schedule_core.cpp"
