+++ include/handlers/signals/signal_listener.h
@@ -2,7 +2,7 @@
 
 #include "event.h"
 #include "signal_constants.h"
-#include "signal_listener_impl_base.h"
+#include "signals/signal_listener_impl_base.h"
 #include "window.h"
 #include <memory>
