+++ tests/filters/test_filter_group.h
@@ -2,7 +2,7 @@
 #include <catch2/reporters/catch_reporter_event_listener.hpp>
 #include <catch2/reporters/catch_reporter_registrars.hpp>
 
-#include "filter_group.h"
+#include "filters/filter_group.h"
 
 TEST_CASE("FilterGroup") {
     FilterGroup filterGroup = FilterGroup();
